// src/gen-ts/src/models/MaterialUnit.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const MaterialUnit = {
    SAC: 'SAC',
    L: 'L',
    KG: 'KG',
    M2: 'M2',
    M3: 'M3',
    KIT: 'KIT',
    POT: 'POT',
    PNL: 'PNL',
    FEU: 'FEU',
    BAR: 'BAR',
    T: 'T',
    M: 'M',
    FFT: 'FFT',
    U: 'U'
} as const;
export type MaterialUnit = typeof MaterialUnit[keyof typeof MaterialUnit];


export function instanceOfMaterialUnit(value: any): boolean {
    for (const key in MaterialUnit) {
        if (Object.prototype.hasOwnProperty.call(MaterialUnit, key)) {
            if (MaterialUnit[key as keyof typeof MaterialUnit] === value) {
                return true;
            }
        }
    }
    return false;
}

export function MaterialUnitFromJSON(json: any): MaterialUnit {
    return MaterialUnitFromJSONTyped(json, false);
}

export function MaterialUnitFromJSONTyped(json: any, ignoreDiscriminator: boolean): MaterialUnit {
    return json as MaterialUnit;
}

export function MaterialUnitToJSON(value?: MaterialUnit | null): any {
    return value as any;
}

export function MaterialUnitToJSONTyped(value: any, ignoreDiscriminator: boolean): MaterialUnit {
    return value as MaterialUnit;
}